Began work on new config
Players suggested I do SOMETHING with the config, the main config is a
bit messy.

I am beginning to break the configs up per mob... It should be much
easier to manage on my end and the player's end when something goes
wrong. So far only Zombie has been done. :)